refactor: extract releases
